chore: remove references to "experimental" analysis
