+++ lib/index.ts
@@ -1,11 +1,10 @@
 import { display } from "./display";
 import * as dockerFile from "./docker-file";
-import { experimentalAnalysis } from "./experimental";
+import { scan } from "./scan";
 import {
   ContainerTarget,
   Fact,
   Identity,
-  PluginOptions,
   PluginResponse,
   ScanResult,
 } from "./types";
@@ -20,21 +19,3 @@ export {
   Identity,
   Fact,
 };
-
-async function scan(options?: Partial<PluginOptions>): Promise<PluginResponse> {
-  if (!options) {
-    throw new Error("No plugin options provided");
-  }
-
-  const targetImage = options.path;
-  if (!targetImage) {
-    throw new Error("No image identifier or path provided");
-  }
-
-  const dockerfilePath = options.file;
-  const dockerfileAnalysis = await dockerFile.readDockerfileAndAnalyse(
-    dockerfilePath,
-  );
-
-  return await experimentalAnalysis(targetImage, dockerfileAnalysis, options);
-}
