+++ test/windows/plugin.spec.ts
